Notify user of available animal products in barns
Prior to this change, we were only notifying the user about items that
were available for harvest in the barn (auto grabber, cheese press,
etc). This change adds checks for any livestock that have produce
available for harvesting, and list which tools the user should take with
them that day.